New Stimuli Syntax
Updated RGC_ stimuli to new syntax. All stimuli are now defined by a
parameter dictionary, a 'buildStimulus' function which defines objects,
an 'iterateStimulus' function which defines how the movie is constructed
frame-by-frame, and any auxiliary functions needed for either 'build~'
or 'iterate~'.

The sequence of their execution is then defined in an Ordered Dictionary
'dispatcher', which is then executed. In principle the dispatcher
function will always be the same and could be abstracted to QDS.

All new stimuli have been compiled in QDS to confirm correct operation.