New tool: Probe center
Added functionnalities to probe the receptive field of neurons.
The probe center is an oval that oscilate between a low and high
intensity grey, and that we move with the mouse to detect were the
recorded neuron respond.
Therefore it need a live display.

Modifications:
**QDSpy_GUI_main_ui
- added the tab and the buttons for the probe_center
**QDSpy_GUI_main.py
- added a function OnClick_btnProbeStart() called when the probe button
is pressed
- copied and adapt runStim() for the probe center
- updated keyPressEvent(), closeEvent() and updateAll() to behave like
runStim with the probe center
**QDSpy_core.py
- copied function associated with the PRESENTATION and adapt it for the
PROBE_CENTER
**QDSpy_multiprocessing:
- Added probe_center to the different enumerations.
**QDSpy_probeCenter:
Contain the probeCenter class and a main function to run the probe
center. Inside it I use openGL functions to try to keep the probe center
located on the mouse pointer (it doesn't work every time...)