Addition of marker size in parameters
The markerScrWidthFract property existed previoulsy in the code. I added
this property in the ini file to make it easy to change.
We prefer to have it in a larger size.
Also in QDSpy_stim_draw, the way the vertices are specified, it made the
square going outside of the window. I don't know if it was intentional
(with the overlay) but I fixed this. Up to you to keep it like this or
not. But if you keep it, the ratio is now correct (instead of half the
expected size previously)