Fix probe movement
A previous fix was correcting the size of the probe according to the
scale inside QDSpy. But then the probe was not following the mouse, and
position of the probe was wrong.
Now the movement and the position is correct. The position is given in
pixel.
Also I changed the format of DATA display, to be coherent with the other
DATA logging.