Anti marker display
Where the marker is displayed, the stimulus can also be displayed if its
on the full screen.
So I added an "antimarker", that is by default black with alpha 255. The
color can be change within QDSpy.ini, and if someone want to get rid of
it he can simply put the alpha of the antimarker at 0.
If the option markershowonscreen is False, the marker and antimarker
won't be displayed.