Parameter on GUI_timeout
Some stimuli can be heavy (like a checkerboard at 60Hz for 15min ->
700mb, even with some compression)
If the stimuli is too long to load, QDSpy behaviour is strange. So I
added the GUI_timeout as parameter. By increasing it, we can leave more
time to QDSpy to load a stimulus. Like this no more problem with larger
stimuli.